Archive/Implementation/Front End/video-interview-system/requirements.txt
Streamlit==1.10.0
moviepy==1.0.3
librosa==0.8.1
ffmpeg-python==0.2.0
tensorflow==2.8.0
scikit-learn==1.0.2
openai-whisper==1.0.0